+++ NEWS.md
@@ -1,5 +1,8 @@
 # conflicted 0.1.0.9000
 
+* `conflicted_prefer()` allows you to declare a persistent preference 
+  (within a session) for one function over another (#4)
+
 * Added a `NEWS.md` file to track changes to the package.
 
 * conflicted now listens for `detach()` events and removes conflicts that